+++ CHANGELOG.md
@@ -7,14 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Target enums 
-- `mypy` for targets package
+- `mypy` for targets and intervals
 
 ### Changed
 - Renamed `bounds_transform_func` target attribute to `transformation`
+- `Interval.is_bounded` now implements the mathematical definition of boundedness
 - Moved and renamed target transform utility functions
 
 ### Removed
 - Conda install instructions and version badge
+
+### Deprecations
+- `Interval.is_finite` replaced with `Interval.is_bounded`
 
 ## [0.7.1] - 2023-12-07
 ### Added